+++ db_es/callback.py
@@ -50,104 +50,6 @@ def listofstrings(v):
     return rv
 
 
-DOCMAP = {
-    "iss": [
-        # docname  esname  converter
-        ("comment", "comment", noconversion),
-        ("cycle", "cycle", int),
-        ("detectors", "detectors", noconversion),
-        ("e0", "e0", noconversion),
-        ("edge", "edge", noconversion),
-        ("element", "element", noconversion),
-        ("experiment", "experiment", noconversion),
-        ("group", "group", noconversion),
-        ("name", "name", noconversion),
-        ("num_points", "num_points", noconversion),
-        ("plan_name", "plan_name", noconversion),
-        ("PI", "pi", noconversion),
-        ("PROPOSAL", "proposal", noconversion),
-        ("SAF", "saf", noconversion),
-        ("scan_id", "scan_id", noconversion),
-        ("time", "time", noconversion),
-        ("trajectory_name", "trajectory_name", noconversion),
-        ("uid", "uid", noconversion),
-        ("year", "year", int),
-        ("time", "date", toisoformat),
-    ],
-    "xpd": [
-        # docname  esname  converter
-        ("bt_experimenters", "experimenters", listofstrings),
-        ("bt_piLast", "pi", noconversion),
-        ("bt_safN", "saf", str),
-        ("bt_wavelength", "wavelength", float),
-        ("composition_string", "formula", noconversion),
-        ("dark_frame", "dark_frame", bool),
-        ("group", "group", noconversion),
-        ("lead_experimenter", "pi", noconversion),
-        ("notes", "comment", noconversion),
-        ("num_points", "num_points", noconversion),
-        ("plan_name", "plan_name", noconversion),
-        ("sample_composition", "composition", normalize_counts),
-        ("scan_id", "scan_id", noconversion),
-        ("sp_computed_exposure", "sp_computed_exposure", float),
-        ("sp_num_frames", "sp_num_frames", int),
-        ("sp_plan_name", "sp_plan_name", noconversion),
-        ("sp_time_per_frame", "sp_time_per_frame", float),
-        ("sp_type", "sp_type", noconversion),
-        ("time", "time", noconversion),
-        ("time", "date", toisoformat),
-        ("uid", "uid", noconversion),
-        ("time", "year", lambda t: int(toisoformat(t)[:4])),
-    ],
-}
-
-
-def iss_filter(x):
-    return True
-
-
-xpd_pis = (
-    "0713_test",
-    "Abeykoon",
-    "Antonaropoulos",
-    "Assefa",
-    "Banerjee",
-    "Benjiamin",
-    "Billinge",
-    "Bordet",
-    "Bozin",
-    "Demo",
-    "Dooryhee",
-    "Frandsen",
-    "Ghose",
-    "Hanson",
-    "Milinda and Runze",
-    "Milinda",
-    "Pinero",
-    "Robinson",
-    "Sanjit",
-    "Shi",
-    "Test",
-    "Yang",
-    "billinge",
-    "simulation",
-    "test",
-    "testPI",
-    "testPI_2",
-    "testTake2",
-    "xpdAcq_realase",
-)
-
-
-def xpd_filter(x):
-    if "bt_piLast" not in x:
-        return True
-    elif x["bt_piLast"] in xpd_pis:
-        return True
-    else:
-        return False
-
-
 def esdocument(docmap, entry):
     rv = {}
     for doc_key, esname, fcnv in docmap:
