+++ db_es/tests/test_callback.py
@@ -1,7 +1,7 @@
 import time
 import pytest
 
-from db_es.callback import ElasticInsert, DOCMAP, xpd_filter, iss_filter
+from db_es.callback import ElasticInsert, noconversion
 
 # TODO: fill these in with prototype data!
 xpd_doc = {"bl": "xpd", 'uid': 'hi'}
@@ -9,14 +9,55 @@ xpd_bad_doc = {"bt_piLast": "not Simon", "bl": "xpd", 'uid': 'hi1'}
 iss_doc = {"bl": "iss", 'uid': 'hi2'}
 
 
+xpd_pis = (
+    "0713_test",
+    "Abeykoon",
+    "Antonaropoulos",
+    "Assefa",
+    "Banerjee",
+    "Benjiamin",
+    "Billinge",
+    "Bordet",
+    "Bozin",
+    "Demo",
+    "Dooryhee",
+    "Frandsen",
+    "Ghose",
+    "Hanson",
+    "Milinda and Runze",
+    "Milinda",
+    "Pinero",
+    "Robinson",
+    "Sanjit",
+    "Shi",
+    "Test",
+    "Yang",
+    "billinge",
+    "simulation",
+    "test",
+    "testPI",
+    "testPI_2",
+    "testTake2",
+    "xpdAcq_realase",
+)
+
+
+def xpd_filter(x):
+    if "bt_piLast" not in x:
+        return True
+    elif x["bt_piLast"] in xpd_pis:
+        return True
+    else:
+        return False
+
 @pytest.mark.parametrize(
     "idx, dm, bl, f, doc",
     zip(
         ["xpd", "iss"],
-        [DOCMAP["xpd"],
-         DOCMAP["iss"]],
+        [[("uid", "uid", noconversion), ('bl', 'bl', noconversion)],
+         [("uid", "uid", noconversion), ('bl', 'bl', noconversion)]],
         ["xpd", "iss"],
-        [xpd_filter, iss_filter],
+        [xpd_filter, lambda x: True],
         [xpd_doc, iss_doc],
     ),
 )
@@ -33,7 +74,7 @@ def test_no_op_callback(es):
     cb = ElasticInsert(
         es=es,
         esindex="bad_xpd",
-        docmap=DOCMAP["xpd"],
+        docmap=[("uid", "uid", noconversion), ('bl', 'bl', noconversion)],
         beamline="xpd",
         criteria=xpd_filter,
     )
